Create order (#5)
* readme add logo, add image field to entity

* update README.md

* add variant to order_details, remove unit_cost, order response cleaner

* clean imports

* update model to create orders

* update model and references between ordere detail and extra detail

* organize imports

* organize imports, clean repos, createOrder OK, clean code

+++ src/main/java/com/mdmc/posofmyheart/domain/models/Category.java
@@ -5,7 +5,7 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Size;
 
 public record Category(
-        Integer idCategory,  // Null para categorías nuevas
+        Long idCategory,  // Null para categorías nuevas
 
         @NotBlank(message = "El nombre de categoría es obligatorio")
         @Size(max = 50, message = "El nombre no puede exceder 50 caracteres")
